Remove remainder of tests, and rename Difficulty class to ProofOfWork
