+++ examples/decode.txout.proof.php
@@ -1,6 +1,6 @@
 <?php
 
-require "vendor/autoload.php";
+require "../vendor/autoload.php";
 
 $proof = '0100000090f0a9f110702f808219ebea1173056042a714bad51b916cb6800000000000005275289558f51c9966699404ae2294730c3c9f9bda53523ce50e9b95e558da2fdb261b4d4c86041b1ab1bf930900000002bc56aae9c0b9a19d49250c9bf9bf90b3c1ee3ac9096410a1eb179e1e92f90a66201f4587ec86b58297edc2dd32d6fcd998aa794308aac802a8af3be0e081d674013d';
 
@@ -13,7 +13,6 @@ $tree = $filtered->getPartialTree();
 
 $hashes = $tree->getHashes();
 $matches = [];
-//$tree->extractMatches($matches);
 
 echo
     " Block Information: " . PHP_EOL .
