+++ examples/signedmessage.verify.php
@@ -4,8 +4,8 @@ require "../vendor/autoload.php";
 
 use BitWasp\Bitcoin\Bitcoin;
 use BitWasp\Bitcoin\MessageSigner\MessageSigner;
-use BitWasp\Bitcoin\Serializer\MessageSigner\SignedMessageSerializer;
-use BitWasp\Bitcoin\Crypto\EcAdapter\Impl\PhpEcc\Serializer\Signature\CompactSignatureSerializer;
+use BitWasp\Bitcoin\Address\PayToPubKeyHashAddress;
+use BitWasp\Bitcoin\Crypto\EcAdapter\Serializer\Signature\CompactSignatureSerializerInterface;
 
 Bitcoin::setNetwork(\BitWasp\Bitcoin\Network\NetworkFactory::bitcoinTestnet());
 
@@ -19,8 +19,13 @@ IBpGR29vEbbl4kmpK0fcDsT75GPeH2dg5O199D3iIkS3VcDoQahJMGJEDozXot8JGULWjN9Llq79aF+F
 
 $ec = Bitcoin::getEcAdapter();
 
+/** @var PayToPubKeyHashAddress $addr */
 $addr = \BitWasp\Bitcoin\Address\AddressFactory::fromString($address);
-$serializer = new SignedMessageSerializer(new CompactSignatureSerializer(Bitcoin::getMath()));
+
+/** @var CompactSignatureSerializerInterface $cs */
+$cs = \BitWasp\Bitcoin\Crypto\EcAdapter\EcSerializer::getSerializer($ec, CompactSignatureSerializerInterface::class);
+$serializer = new \BitWasp\Bitcoin\Serializer\MessageSigner\SignedMessageSerializer($cs);
+
 $signedMessage = $serializer->parse($sig);
 
 $signer = new MessageSigner($ec);
