+++ src/Chain/BlockLocator.php
@@ -62,41 +62,4 @@ class BlockLocator extends Serializable
     {
         return (new BlockLocatorSerializer())->serialize($this);
     }
-
-    /**
-     * @param int $height
-     * @param BlockIndex $index
-     * @param bool $all
-     * @return BlockLocator
-     */
-    public static function create($height, BlockIndex $index, $all = false)
-    {
-        $step = 1;
-        $hashes = [];
-        $pIndex = $index->hash()->fetch($height);
-
-        while (true) {
-            array_push($hashes, Buffer::hex($pIndex, 32));
-            if ($height == 0) {
-                break;
-            }
-
-            $height = max($height - $step, 0);
-            $pIndex = $index->hash()->fetch($height);
-            if (count($hashes) >= 10) {
-                $step *= 2;
-            }
-        }
-
-        if ($all || count($hashes) == 1) {
-            $hashStop = Buffer::hex('00', 32);
-        } else {
-            $hashStop = array_pop($hashes);
-        }
-
-        return new self(
-            $hashes,
-            $hashStop
-        );
-    }
 }
